+++ submissions/chimoney-payout-airtime/package.json
@@ -13,6 +13,7 @@
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.6",
     "sweetalert2": "^11.4.38",
+    "validator": "^13.7.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
